requirements.txt
aiofiles==22.1.0
aiosqlite==0.19.0
anyio==3.6.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asn1crypto==1.5.1
astroid==2.12.12
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
autopep8==2.0.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==6.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.7
Django==3.2
django-ckeditor==6.5.1
django-extensions==3.2.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.0
executing==1.2.0
Faker==18.4.0
fastjsonschema==2.16.3
filelock==3.12.2
Fiona==1.9.2
fonttools==4.38.0
fqdn==1.5.1
geopandas==0.12.2
graphviz==0.20.1
h11==0.14.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
isort==5.10.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
lxml==5.1.0
markdownify==0.11.6
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==2.0.5
munch==2.5.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.1
oscrypto==1.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.8.0
parso==0.8.3
pdfkit==1.0.0
pickleshare==0.7.5
Pillow==9.3.0
pipenv==2023.7.23
pipx==1.1.0
platformdirs==3.10.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pyChatGPT==0.4.0
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.15.0
pylint==2.15.5
PyNaCl==1.5.0
pyparsing==3.0.9
pyproj==3.5.0
pyrsistent==0.19.3
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.0
python-json-logger==2.0.7
python-owasp-zap-v2.4==0.0.21
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==3.0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
selenium==4.7.2
Send2Trash==1.8.0
shapely==2.0.1
six==1.15.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
stack-data==0.6.2
stripe==5.0.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.9.0
tzdata==2022.6
undetected-chromedriver==3.1.7
uri-template==1.2.0
uritools==4.0.0
urllib3==1.26.12
userpath==1.8.0
virtualenv==20.24.3
virtualenv-clone==0.5.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
wrapt==1.14.1
wsproto==1.2.0
y-py==0.5.9
ypy-websocket==0.8.2
